feat: 네비게이션 관련 파일 타입스크립트로 변경
